Add View->Hide Unused Fields
This feature allows the user to hide elements not related to the current
memory channel's configuration.  For example, in "Tone" mode, the ToneSql,
DTCS, etc fields are not shown.  If Duplex is (None) then the offset field
is blank.  This should streamline operation of CHIRP with a screen reader
and may be preferable to some sighted hams as well.

Fixes ticket:92