Improve sub-device treatment in the UI
This changes how CHIRP exposes devices with sub-devices to the user, moving
from an strange per-sub-device top-level tab to minor tabs along the left.
This means replicating the memories, bank, and bank names tabs once per
sub-device. In order to prevent two radiothreads from trying to talk to
the same live device at the same time, RadioThread is extended to take a
parent thread, which it shares the same runlock with for exclusion.

Fixes #739