[bf888] Standardize cloning routines and change name
This patch makes the BF888/H777 driver pass the functional tests. Most
of the change is around making the clone routines return conventional
error messages and the proper exceptions for errors. Also, to run the
functional tests, we have to have a predictable VENDOR and MODEL for
the driver. The chinese name is cool (and correct) but it's going to
get in the way of testing on some systems. For now, call this driver
the Baofeng BF-888. Later, if we can figure out how to differentiate
files for them, we can have a subclass for the H777.

Related to #701