[UV-5R] Update Firmware Version Detection
Fix firmware version block length (14 characters instead of 16). The
firmware version check was failing on characters outside of the firmware
version block.

Swap the check order for BFB and BFS in the firmware version string so
the first occurance will always be selected. The previous method was
allowing different firmware versions to be detected as being the same.

Related to bug #1751