Suppress startup debugging messages (#2347)
This patch changes the verbose startup messages to use LOG.info calls.
This change is required to make the CLI remotely usable, as it's
unreasonable to print so much information in that context.

I kept this separate from the last patch, because I can imagine that
this might impact the automated/buildbot testing.  To avoid regressions,
the previous output can be generated by setting 'CHIRP_DEBUG=info' in the
environment.