[GT-890] Remove direct support for the Sainsonic GT-890, fixes #3533
This patch is to remove the direct support for the Sainsonic GT-890 as it's
a perfect clone of the QYT KT8900, this applies to the two known variants of
the GT-890 so far.

We are paving the way for the new radio.ALIASES feature that will put aliases
on the UI in a user friendly way. We have internal tests going on about it.

73 Pavel CO7WT.