[ui] Deprecate .chirp file format by disabling export. #3689
The .chirp file format provides no additional features compared to .CSV and is
therefore redundant. .chirp is the more complex and less frequently used format
of the two, so to eliminate the redundancy it should be deprecated.

This patch removes the ability to export as .chirp, leaving no way to generate
new .chirp files. It preserves the ability to open and import .chirp files, so
any existing .chirp files can still be opened, edited, and exported to other
formats.