[PATCH][Feature] Add the know clones of the btech with the new alias feature, fixes #3749
This patch explicitly adds the following Btech clones to the UI:
* Jetstream JT2705M
* Juentai JT-6188 Mini
* Juentai JT-6188 Plus
* Sainsonic GT-890

This radios do show properly now in the download/upload dialog as vendor/model, this will ease the handling of clones in the user side.

But it has a catch and the users will be notified on the chirp_user list: when you opens a saved image of a clone it will show in Chirp as the real radio, not the clone.

73 Pavel CO7WT