Add support for Alinco DJ-G7EG. Partially addresses #699
This patch adds basic support for the DJ-G7EG. The patch was tested on
Linux and Windows 7 and can reliably download memories from, and upload
memories to the radio. Other flavours of the DJ-G7 are not supported
yet because I don't know whether they use the same memory map or not.

The patch introduces a new class to the alinco.py driver. This is some-
what unfortunate because the radios supported by alinco.py so far are
all quite similar. The DJ-G7EG is considerably more different from the
others, yet still sufficiently similar to re-use some of the methods
implemented in AlincoStyleRadio. I'll leave it up to the reviewers to
comment on that. Code that might be of particular interest to reviewers
is marked with "FIXME" comments.

The patch does not support changing bank assignments, groups, and/or
radio settings. I am planning to add dynamic bank assignements as soon
as I have figured out how to do this best with Chirp.