Add support for xdg base directories
Modern Linux distributions use the XDG Base Directory Specification,
which decides where various files should be stored. It is possible to
respect this specification by using glib (GLib).

This adds glib as a new dependency, however it should already be installed
as a part of pygobject 2.28.x, unless a distribution splits upstream
packages.

Example of where files are stored on a typical Linux desktop:

- Configs: ~/.config/chirp/
- Data: ~/.local/share/chirp/

Example of where this differs, with Flatpak:

- Configs: ~/.var/app/com.danplanet.chirp/config/chirp/
- Data: ~/.var/app/com.danplanet.chirp/data/chirp/

These changes ensure any of these (and any other) will work.